Divide JS code into modules
